+++ src/Sharkable/SharkEndpoint/Extensions/EndpointRouteBuilderExtension.cs
@@ -17,7 +17,7 @@ public static class EndpointRouteBuilderExtension
     public static IEndpointConventionBuilder SharkMapGet(this IEndpointRouteBuilder app, [StringSyntax("Route")]string sharkPattern, Delegate sharkHandler)
     {
         var p = sharkPattern.GetCaseFormat(Shark.SharkOption.Format) ?? 
-            throw new Exception("error when formatting given string path");
+            throw new InvalidOperationException("error when formatting given string path");
         
         return app.MapGet(p, sharkHandler);
     }